requirements.txt

# Core dependencies
pillow==10.2.0
pypng==0.20220715.0
python-dotenv==1.0.1
qrcode==7.4.2
typing_extensions==4.10.0
validators==0.23.2

# Development dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-pylint==0.21.0
black==23.12.1
flake8==6.1.0
mypy==1.8.0
coverage==7.3.2
